Update: Post 식별자 타입 변경 (Int -> Long)

+++ src/main/kotlin/kea/dpang/faq/controller/PostController.kt
@@ -33,7 +33,7 @@ class PostController(private val postService: PostService) {
 
     @GetMapping("/{postId}")
     fun readPost(
-        @PathVariable postId: Int
+        @PathVariable postId: Long
     ): ResponseEntity<SuccessResponse<PostResponse>> {
 
         val post = postService.getPost(postId)
