+++ src/main/kotlin/kea/dpang/faq/exception/PostNotFoundException.kt
@@ -2,4 +2,4 @@ package kea.dpang.faq.exception
 
 import java.lang.RuntimeException
 
-class PostNotFoundException(postId: Int) : RuntimeException("해당 ID의 게시글을 찾을 수 없습니다. ID: $postId")
+class PostNotFoundException(postId: Long) : RuntimeException("해당 ID의 게시글을 찾을 수 없습니다. ID: $postId")
